step 1 move existing types to v7
