+++ types/react-relay/v7/lib/relay-experimental/fetchQuery.d.ts
