+++ types/react-relay/v7/lib/relay-experimental/useEntryPointLoader.d.ts
